Represent angle in ScheduleManeuver event as f64
It's more convenient that way, if Frame is going to contain structs from
shared::game.

+++ source/rust/client/src/lib/interface.rs
@@ -15,7 +15,7 @@ pub enum InputEvent {
 	StartBroadcast(String),
 	StopBroadcast,
 
-	ScheduleManeuver(f32),
+	ScheduleManeuver(f64),
 
 	Quit,
 }
